feat: 수의시설 목록 조회 요청, 응답 DTO 구현

+++ src/main/java/yong/petdoc/repository/vetfacility/VetFacilityCustomRepository.java
@@ -1,6 +1,5 @@
 package yong.petdoc.repository.vetfacility;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 import yong.petdoc.domain.vetfacility.VetFacility;
@@ -9,8 +8,8 @@ import yong.petdoc.domain.vetfacility.VetFacilityType;
 public interface VetFacilityCustomRepository {
 
 	List<VetFacility> findVetFacilities(
-		BigDecimal latitude,
-		BigDecimal longitude,
+		Double latitude,
+		Double longitude,
 		Integer radius,
 		VetFacilityType type
 	);
